BS_MAP_BOT-main/Bs_Map-main/requirements.txt
disnake
Pillow
python-dotenv
numpy